feat: modify docker-entrypoint
1. check application directory
2. check project not create

+++ conf/docker/dotnet.3.sdk/docker-entrypoint.sh
@@ -3,19 +3,33 @@
 
 set -e
 
-if [ "$1" = "dotnet" ]; then
-    if [ "$2" = "publish" ]; then
-        cd /app
-        rm -rf publish/*
-        dotnet publish --configuration Release -o publish
-    fi
-    if [ -e /app/publish/Service.dll ];
+DOTNET_APP_PATH=/app
+DEVELOP_MODE=
+
+if [ "$1" = "dotnet" ] && [ -e ${DOTNET_APP_PATH} ]; then
+    cd ${DOTNET_APP_PATH}
+    if [ $( find . -type f -iname "*.sln" | wc -l ) -gt 0 ];
     then
-        dotnet /app/publish/Service.dll
+        if [ "$2" = "publish" ]; then
+            rm -rf publish/*
+            dotnet publish --configuration Release -o publish
+        fi
+        if [ -e /app/publish/Service.dll ];
+        then
+            dotnet /app/publish/Service.dll
+        else
+            echo "$0: project not publish."
+            DEVELOP_MODE=1
+        fi
     else
-        echo "$0: project not publish."
-        tail -f /dev/null
+        echo "$0: project not create."
+        DEVELOP_MODE=1
     fi
+else
+    echo "$0: .NET application folder not find."
+    exit 1
 fi
+
+[ ! -z ${DEVELOP_MODE} ] && tail -f /dev/null
 
 exec "$@"
